Supervisor script
Co-authored-by: GuilhermeSantosGabriel <guilhermesangabriel@gmail.com>
Script that controls the simulation environment, randomly moving obstacles as the drone passes them. After creating the "env_supervisor_controller" in the Wizards menu, it is necessary to create a supervisor node so as to run the script in Webots, setting its supervisor field to "true" and selecting the corresponding controller in the controller field.